Fix up node name resolution from within Pod
At present the hostname resolution is handed back the Pod name rather
than the node name when run within a Kubernetes deployment. We fix
this up by specifying a NODE_NAME environment variable and passing
it in - falling back to hostname lookup if it's unset, as in e.g.
node-local testing outside of a deployment context.